+++ Runtime/Graphics/Shaders/CLineRendererShadersHLSL.cpp
@@ -110,7 +110,7 @@ struct HLSLLineDataBindingFactory : CLineRendererShaders::IDataBindingFactory
 
         boo::IGraphicsBuffer* uniforms[] = {ubufInfo.first};
         boo::PipelineStage stages[] = {boo::PipelineStage::Vertex};
-        size_t ubufOffs[] = {ubufInfo.second};
+        size_t ubufOffs[] = {size_t(ubufInfo.second)};
         size_t ubufSizes[] = {sizeof(CLineRenderer::SDrawUniform)};
 
         renderer.m_shaderBind = ctx.newShaderDataBinding(pipeline, nullptr, vbufInfo.first,
